Add script and workflow to test code-scanning-rubocop
The rubocop.yml workflow here is the one I added in
https://github.com/EliahKagan/sieve-ksh/pull/4, including
modifications made as part of that PR, but excluding later changes.

The idea here is to produce an error in the upload-sarif step when
the workflow is run on a Dependabot branch with the push trigger
(due to a Dependabot PR being opened, or changes made to it). Then
a more robust fix than the changes I made to the workflow by the
commits I added to https://github.com/EliahKagan/sieve-ksh/pull/5
can be found.